Move TNotification and IJobApplicationNotification to interfaces directory
