+++ src/interfaces/Notification/index.ts
@@ -1,0 +1,13 @@
+import { JOB_APPLICATION_NOTIFICATION_TYPE } from "$typenames";
+import { IJobApplication } from "$interfaces/JobApplication";
+
+export type TNotification = IJobApplicationNotification;
+
+export interface IJobApplicationNotification {
+  __typename: JOB_APPLICATION_NOTIFICATION_TYPE;
+  uuid: string;
+  message: string;
+  createdAt: string;
+  adminEmail: string;
+  jobApplication: IJobApplication;
+}
