+++ src/models/hooks/queries/useNotifications.ts
@@ -1,7 +1,6 @@
 import { GET_NOTIFICATIONS } from "$queries";
 import { usePaginatedQuery } from "$hooks";
-import { IJobApplication } from "$interfaces/JobApplication";
-import { JOB_APPLICATION_NOTIFICATION_TYPE } from "$typenames";
+import { TNotification } from "$interfaces/Notification";
 
 export const useNotifications = () =>
   usePaginatedQuery<{}, TNotification>({
@@ -10,14 +9,3 @@ export const useNotifications = () =>
     variables: {},
     timestampKey: "createdAt"
   });
-
-export type TNotification = IJobApplicationNotification;
-
-export interface IJobApplicationNotification {
-  __typename: JOB_APPLICATION_NOTIFICATION_TYPE;
-  uuid: string;
-  message: string;
-  createdAt: string;
-  adminEmail: string;
-  jobApplication: IJobApplication;
-}
